Migração para BigDecimal e consolidação do domínio bancário

+++ src/service/LedgerService.java
@@ -1,4 +1,3 @@
-// src/service/LedgerService.java
 package service;
 
 import domain.account.Account;
@@ -6,14 +5,8 @@ import exceptions.AccountNotFoundException;
 import infrastructure.logging.AuditLogger;
 import repository.AccountRepository;
 
-/**
- * Serviço de ledger (histórico transacional).
- * Registra movimentações e fornece extrato.
- * 
- * Capítulos:
- * 7, 16 - Coleções (GenericLinkedList)
- * 14 - Strings (formatação)
- */
+import java.math.BigDecimal;
+
 public class LedgerService {
     private AccountRepository accountRepo;
     private AuditLogger logger;
@@ -25,18 +18,14 @@ public class LedgerService {
 
     /**
      * Registra uma transação genérica na conta.
-     * Assume que Account possui método recordTransaction.
      */
-    public void record(String accountNumber, String description, double amount)
+    public void record(String accountNumber, String description, BigDecimal amount)
             throws AccountNotFoundException {
         Account account = accountRepo.findById(accountNumber);
         account.recordTransaction(description, amount);
         logger.audit("Transação na conta " + accountNumber + ": " + description + " R$ " + amount);
     }
 
-    /**
-     * Obtém o extrato formatado da conta.
-     */
     public String getStatement(String accountNumber) throws AccountNotFoundException {
         Account account = accountRepo.findById(accountNumber);
         return account.getStatement();
